Refactor: Make PluginExtensionsAndVersionValidator implement PluginPostLoadHook
